fixed reset error
reapplied "[]" fix since it would reset users favorites back to empty when ever a new search was made. Fixes this issue but once again cannot remove the final saved favorite due to bandaid fix applied. Able to pull and log the documents promise object and view user id. Can be seen in console each time a search is made. Just need to parse the string out and utilize it to update favorites list so leaving and navigating back will maintain the favorites list.